requirements.txt
django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
pymongo>=4.13.0,<5.0.0
django-cors-headers>=4.3.0,<5.0.0
python-dotenv>=1.1.0,<2.0.0
pytz>=2024.1
drf-spectacular>=0.27.0,<0.28.0
drf-spectacular-sidecar>=2024.3.4,<2025.0.0
sqlparse>=0.3.1
